Improved memory allocation in ap_phot
Photutil's make_source_mask uses a runaway amount of memory and eventually causes ap_phot.py to crash. Removed it from the background estimation procedure, so will have to worry about masking sources in background annuli in the future. For now, background estimate takes all values in the annulus, drops any that are negative (to deal with references near the edge of the detector), does a 3-sigma clipping on the remaining values, and takes the mean of the sigma-clipped distribution as the per-pixel sky background estimate.