Added "live plot" mode
Setting live_plot=True in the call to fixed_circular_aperture_photometry in ap_phot.py will now display a "live plot" of the analysis as it progresses. It will show the entire field, a cutout surrounding the target, time series of normalized target/ALC flux, and corrected target flux.