Added ap_range
Added a function called ap_range() to ap_phot.py. It measures the x/y FWHM seeing of the target across a set of images and determines a reasonable range of aperture sizes based on those measurements. It writes out this range to a file called "aperture_range.csv" in the /data/tierras/lightcurves/DATE/TARGET/FFNAME/ directory.