Added variable aperture photometry
Changed fixed_circular_aperture_photometry() to circular_aperture_photometry(). circular_aperture_photometry takes an argument called 'type' which can be set to 'fixed' or 'variable'. If 'variable', it will used measurements of the target's FWHM (created with ap_range()) to create a time series of aperture radii. If 'variable', the ap_radii argument is treated as a multiplicative factor of this FWHM time series.