Added padding to bpm, added bkg plot
Added extra padding to the bad pixel mask used for detecting sources/identifying reference stars. The two bad columns in the upper/lower halves now both have +/-20 pixels of padding to either side. Also added +/-15 pixels of padding to the mask around the divide between the upper/lower halves. This should help with avoiding selecting references that fall too close to these features. 

Added a histogram of background annulus pixel values to the live plot. It shows the full annulus distribution, the sigma-clipped distribution, and the value being measured as the background estimate.