Fixed FWHM measurements
Adopted a convention where 'X' is the major axis of the 2D Gaussian used to measure FWHM seeing, and 'Y' is the minor axis. Prior versions just let the code determine which axis was 'X' and which was 'Y', and this would result in the measurements 'sloshing' between X and Y, since a 2D Gaussian is symmetric under rotation.